Updated script so we only download the jar file once.

+++ discovery_client_gen.py
@@ -9,9 +9,7 @@
 #
 
 from pathlib import Path
-import json
 import os
-import yaml
 import requests
 import sys
 
@@ -20,20 +18,25 @@ OPENAPI_VERSION = "5.4.0"
 CLIENT_LANG = "python"
 
 
-def gen_client(client_dir, openapi_url):
-    Path(client_dir).mkdir(parents=True, exist_ok=False)
+def gen_client(base_dir, client_name, openapi_url):
+    Path(base_dir).mkdir(parents=True, exist_ok=True)
     print("------------------------------------------")
-    print("Client directory: ", client_dir)
+    print("Base directory: ", base_dir)
 
     source_url = f"{OPENAPI_SOURCE}/{OPENAPI_VERSION}/openapi-generator-cli-{OPENAPI_VERSION}.jar"
 
-    jar_path = f"{client_dir}/openapi-generator-cli-{OPENAPI_VERSION}.jar"
-    resp = requests.get(source_url)
-    if resp.status_code != 200:
-        raise Exception("Failed to download " + source_url)
-    with open(jar_path, "wb") as jf:
-        jf.write(resp.content)
-    print("Downloaded ", source_url)
+    jar_path = f"{base_dir}/openapi-generator-cli-{OPENAPI_VERSION}.jar"
+    if not Path(jar_path).exists():
+        resp = requests.get(source_url)
+        if resp.status_code != 200:
+            raise Exception("Failed to download " + source_url)
+        with open(jar_path, "wb") as jf:
+            jf.write(resp.content)
+        print("Downloaded ", source_url)
+
+    client_dir = f"{base_dir}/{client_name}"
+    Path(client_dir).mkdir(parents=True, exist_ok=False)
+    print("Client directory: ", client_dir)
 
     resp = requests.get(openapi_url)
     if resp.status_code != 200:
@@ -43,33 +46,23 @@ def gen_client(client_dir, openapi_url):
         of.write(resp.content)
     print("Downloaded ", openapi_url)
 
-    with open(openapi_file) as jf:
-        openapi_json = json.load(jf)
-
-    openapi_yaml_file = f"{client_dir}/openapi.yaml"
-    with open(openapi_yaml_file, "w") as yf:
-        yaml.safe_dump(openapi_json, yf, default_flow_style=False)
-
     print("")
     print("Using OpenAPI Generator CLI Jar: ", jar_path)
     print("OpenAPI 3.0 Specification File:  ", openapi_file)
-    print("OpenAPI 3.0 Yaml File:           ", openapi_yaml_file)
 
     print("")
     print(f"Generating API {CLIENT_LANG} Client")
 
     python_client_dir = f"{client_dir}/Client"
     Path(python_client_dir).mkdir(parents=True, exist_ok=False)
-    # cmd = f"java -jar {jar_path} generate -i {openapi_yaml_file} -c {generator_config} -g {CLIENT_LANG} -o {python_client_dir}"
-    cmd = f"java -jar {jar_path} generate -i {openapi_yaml_file} -g {CLIENT_LANG} -o {python_client_dir}"
+    cmd = f"java -jar {jar_path} generate -i {openapi_file} -g {CLIENT_LANG} -o {python_client_dir}"
     rc = os.system(cmd)
     print("java exit code: ", rc)
-  
 
 if __name__ == "__main__":
     if len(sys.argv) != 2:
         raise Exception("Usage: discovery_client_gen.py target_directory")
-    client_dirs = sys.argv[1]
-    gen_client(f"{client_dirs}/ingress", "https://console.redhat.com/api/ingress/v1/openapi.json")
-    gen_client(f"{client_dirs}/rhsm", "https://console.redhat.com/api/rhsm/v2/openapi.json")
-    gen_client(f"{client_dirs}/inventory", "https://console.redhat.com/api/inventory/v1/openapi.json")
+    client_dir = sys.argv[1]
+    gen_client(client_dir, "ingress", "https://console.redhat.com/api/ingress/v1/openapi.json")
+    gen_client(client_dir, "rhsm", "https://console.redhat.com/api/rhsm/v2/openapi.json")
+    gen_client(client_dir, "inventory", "https://console.redhat.com/api/inventory/v1/openapi.json")
